+++ h/storage.py
@@ -94,8 +94,8 @@ def create_annotation(request, data, group_service):
     :param data: a dictionary of annotation properties
     :type data: dict
 
-    :param group_service: a service object that adheres to ``memex.interfaces.IGroupService``
-    :type group_service: memex.interfaces.IGroupService
+    :param group_service: a service object that adheres to ``h.interfaces.IGroupService``
+    :type group_service: h.interfaces.IGroupService
 
     :returns: the created and flushed annotation
     :rtype: dict
