+++ tests/h/storage_test.py
@@ -268,7 +268,7 @@ class TestCreateAnnotation(object):
     @pytest.fixture
     def group_service(self, pyramid_config):
         group_service = mock.Mock(spec_set=['find'])
-        pyramid_config.register_service(group_service, iface='memex.interfaces.IGroupService')
+        pyramid_config.register_service(group_service, iface='h.interfaces.IGroupService')
         return group_service
 
     def annotation_data(self):
